Generate on a blank new world.
'New World' folder contains a blank world that will be copied to destination.
Destination, if not given (as second argument), will be created as a folder inside the source's parent,
with the name of the composition.

Remove generator.clear_space, as it is no longer needed. Siginificantly increase speed.
And various other tweaks for code clarity