release 0.2.0
LOTS OF CHANGES:

-------------------------------
1) New build orientation system

With the previous orientation system, whichever option you choose, the long part of the structure must lie on the x-axis.

Now the structure can be generated towards any direction. By the default it's the direction the player is facing,
but this can be specified as a build option using minecraft's rotation syntax.

Because it can be tricky to orient the player just right, or to use the rotation syntax...

-------------------------------
2) User confirmation prompt

The user is informed of the player's location and where the structure will be generated, and prompted to confirm.

It would be wasteful to wait for the user's confirmation while doing nothing.
We actually start generating immediately, just don't save yet, and discard everything immediately if the user denies.

Issue: the generator would crash if the player is inside the world, or enters the world while it's runnning.
This would defeat the point of the confirmation prompt, because the user is supposed to enter the world
to check that everything is correct. Therefore...

-------------------------------
3) Allow generating while the game is running

To avoid the above issue, make a clone of the original world and generate on that one instead.
Upon finishing, replace the original world with the clone one.

This has only been tested to work on Linux and Windows.
Users on other platforms will be prompted to exit the world and not to re-enter until the program finishes.

-------------------------------
4) Verbosity options

This is mostly indended as a convenient way to automatically accept the confirmation (with option -q).
But there are also -qq and -qqq to also hide the progress bar, warnings, and error messages.

-------------------------------

WHAT DID NOT CHANGE: compatibility with music written for version 0.1.3+

WHAT CHANGED FOR THE WORSE: the rotation thing costs around 10% of performance