revise rule: no mixing modes in CompoundSection
When a double-section continuously segues to a single-section or vice-versa,
the single section must either be placed on the left or the right *anyway*.
By not allowing mixed modes, we force the user to choose which side.

This does not apply to non-continuous sections,
because then the single section could be placed in the middle.

Updated JSON schemas.